requirements.txt
# Core ML dependencies
mlflow==2.8.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3

# AWS dependencies
boto3==1.34.0
botocore==1.34.0

# ML and data processing
transformers==4.35.2
torch==2.1.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Development and monitoring
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0

# Security
# bandit==1.7.5
# safety==2.3.5

# API and web
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0